Conflicts:
	Libraries/Mozu.Api.WebToolKit/Mozu.Api.WebToolKit/Logging/ApiLogger.cs
	README.md

+++ SDK/Mozu.Api/Contracts/Core/BehaviorCollection.cs
@@ -15,14 +15,14 @@ using System.Collections.Generic;
 namespace Mozu.Api.Contracts.Core
 {
 		///
-		///	List of application behaviors and details for each record in the list.
+		///	***Always private and should not be published.***
 		///
 		public class BehaviorCollection
 		{
 			///
 			///The number of results listed in the query collection, represented by a signed 64-bit (8-byte) integer. This value is system-supplied and read-only.
 			///
-			public long TotalCount { get; set; }
+			public int TotalCount { get; set; }
 
 			///
 			///An array list of objects in the returned collection.
